feature(exceptions.py): refactor exception to new module
since there are cases we need those in `sdcm.cluster`
and we want to avoid cyclic dependecies
